// src/services/advertise-agency/breadcrumbs-service/src/types/breadcrumbs.ts
// Тип для каждого элемента хлебных крошек (breadcrumbs)
export type BreadcrumbItem = {
    title: string // Заголовок элемента хлебных крошек
    href?: string // Ссылка, на которую ведет данный элемент (не обязательно)
    isCurrent?: boolean // Флаг, указывающий, является ли этот элемент текущим (не обязательно)
}